light: move more fixtures to fixtures.py
I think all of these are framework specific.
